+ Pergamino FBX + UI for current level * Refactor float Highscore Playerprefs to int

+++ Assets/Scripts/ScriptableUpdaters/Templates/SessionData.cs
@@ -7,7 +7,7 @@ public class SessionData : InstantiableScriptable
     public int playerSelected;
     public bool playerAlive = true;
     public Difficulties difficulty = Difficulties.EASY;
-    public float currentScore;
+    public int currentScore;
     public int currentDifficultLevel;
     public PlayerTypes playerPrefab;
